backend/requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0
httpx==0.27.0
pydantic==2.7.1
cachetools==5.3.3
python-dotenv==1.0.1
pytest==8.2.1
pytest-asyncio==0.23.6
respx==0.20.2
anyio==4.3.0
loguru==0.7.2
typing-extensions==4.11.0
